modified:   src/cor_test.c

+++ src/cor_test.c
@@ -397,7 +397,7 @@ SEXP E_test(SEXP _A, SEXP _B, SEXP _W, SEXP _permut, SEXP _threads, SEXP _bidx)
 #pragma omp parallel for num_threads(n_thread)
     for (idx = 0; idx < N_feature; ++idx) {
         // todo: bidx
-        int bidx = REAL(_bidx)[idx] -1;
+        int bidx = INTEGER(_bidx)[idx] -1;
         if (ap[idx] == ap[idx+1] || bp[bidx] == bp[bidx+1]) {
             REAL(LXval)[idx] = 0;
             REAL(LYval)[idx] = 0;
